Make colors consistent regardless of expr ordering
Currently, if you have two graphs which use the same types, the colors
per type might be different, depending on which type comes first.

This makes comparing similar graphs harder visually.

This resolves that by using a sorted map for mapping types to values,
so that if the same set of types are used by two graphs, they will
have the same colors.